fix: update parameter type for splitSegmentDataToChunksAndUploadAsync from Uint8Array to ArrayBuffer

+++ packages/p2p-media-loader-core/src/p2p/peer.ts
@@ -278,7 +278,7 @@ export class Peer {
     this.peerProtocol.sendCommand(command);
     try {
       await this.peerProtocol.splitSegmentDataToChunksAndUploadAsync(
-        data as Uint8Array,
+        data,
         requestId,
       );
       this.sendSegmentDataSendingCompletedCommand(segment, requestId);
